Feature: Create Dao Modal UI APP-227  (#309)
* tranaction context added

* connect modal to deposit page

* fix lint warns

* remove pipeline

* refactore transfers modal and add publish dao

* add pollGas hook

* fix lint error

* add use poll tokens

* fix lint errors

* fix sonar cloud warnings

* fix modal onClose

* add modal to golive page
