+++ packages/web-app/src/containers/goLive/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import styled from 'styled-components';
 import {useTranslation} from 'react-i18next';
 import {useFormContext} from 'react-hook-form';
-import {BigNumberish, ethers} from 'ethers';
 // import {DAOFactory} from 'typechain';
 // TODO reintroduce this by adding back the postInstall script in packages.json
 // that executes the generate-abis-and-types command.
@@ -15,9 +14,7 @@ import Governance from './governance';
 import goLive from 'public/goLive.svg';
 import {useNavigate} from 'react-router-dom';
 import {Dashboard} from 'utils/paths';
-import DAOFactoryABI from 'abis/DAOFactory.json';
-
-import {useProviders} from 'context/providers';
+import {useTransactionContext} from 'context/transactions';
 
 export const GoLiveHeader: React.FC = () => {
   const {t} = useTranslation();
@@ -61,13 +58,7 @@ export const GoLiveFooter: React.FC = () => {
   const {watch} = useFormContext();
   const {reviewCheck} = watch();
   const {t} = useTranslation();
-  const zeroAddress = ethers.constants.AddressZero;
-  const daoDummyName = "Rakesh's Syndicate";
-  const daoDummyMetadata = '0x00000000000000000000000000';
-  const dummyVoteSettings: [BigNumberish, BigNumberish, BigNumberish] = [
-    1, 2, 3,
-  ];
-  const {infura: provider} = useProviders();
+  const {setIsModalOpen} = useTransactionContext();
 
   const IsButtonDisabled = () =>
     !Object.values(reviewCheck).every(v => v === true);
@@ -78,34 +69,7 @@ export const GoLiveFooter: React.FC = () => {
         size="large"
         iconRight={<IconChevronRight />}
         label={t('createDAO.review.button')}
-        onClick={async () => {
-          const contract = new ethers.Contract(
-            '0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0',
-            DAOFactoryABI,
-            provider
-          );
-
-          console.log(
-            'NewDAO Gas:',
-            await contract.estimateGas.newDAO(
-              {
-                name: daoDummyName,
-                metadata: daoDummyMetadata,
-              },
-              {
-                addr: zeroAddress,
-                name: 'TokenName',
-                symbol: 'TokenSymbol',
-              },
-              {
-                receivers: ['0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266'],
-                amounts: [100],
-              },
-              dummyVoteSettings,
-              zeroAddress
-            )
-          );
-        }}
+        onClick={() => setIsModalOpen(true)}
         disabled={IsButtonDisabled()}
       />
     </div>
