+++ packages/web-app/src/context/globalModals.tsx
@@ -12,7 +12,6 @@ type GlobalModalsContextType = {
   isTransferOpen: boolean;
   isTokenOpen: boolean;
   isUtcOpen: boolean;
-  isTransactionOpen: boolean;
   isSelectDaoOpen: boolean;
   isAddActionOpen: boolean;
   isAddressesOpen: boolean;
@@ -24,7 +23,6 @@ type GlobalModalsContextType = {
 type MenuTypes =
   | 'token'
   | 'utc'
-  | 'transaction'
   | 'addAction'
   | 'selectDao'
   | 'default'
@@ -44,8 +42,6 @@ const GlobalModalsProvider: React.FC<Props> = ({children}) => {
     useState<GlobalModalsContextType['isTokenOpen']>(false);
   const [isUtcOpen, setIsUtcOpen] =
     useState<GlobalModalsContextType['isUtcOpen']>(false);
-  const [isTransactionOpen, setIsTransactionOpen] =
-    useState<GlobalModalsContextType['isTransactionOpen']>(false);
   const [isAddActionOpen, setIsAddActionOpen] =
     useState<GlobalModalsContextType['isAddActionOpen']>(false);
   const [isSelectDaoOpen, setIsSelectDaoOpen] =
@@ -62,9 +58,6 @@ const GlobalModalsProvider: React.FC<Props> = ({children}) => {
         break;
       case 'utc':
         setIsUtcOpen(true);
-        break;
-      case 'transaction':
-        setIsTransactionOpen(true);
         break;
       case 'addAction':
         setIsAddActionOpen(true);
@@ -91,9 +84,6 @@ const GlobalModalsProvider: React.FC<Props> = ({children}) => {
         break;
       case 'utc':
         setIsUtcOpen(false);
-        break;
-      case 'transaction':
-        setIsTransactionOpen(false);
         break;
       case 'addAction':
         setIsAddActionOpen(false);
@@ -127,7 +117,6 @@ const GlobalModalsProvider: React.FC<Props> = ({children}) => {
       isTransferOpen,
       isTokenOpen,
       isUtcOpen,
-      isTransactionOpen,
       isAddActionOpen,
       isSelectDaoOpen,
       isAddressesOpen,
@@ -139,7 +128,6 @@ const GlobalModalsProvider: React.FC<Props> = ({children}) => {
       isTransferOpen,
       isTokenOpen,
       isUtcOpen,
-      isTransactionOpen,
       isAddActionOpen,
       isSelectDaoOpen,
       isAddressesOpen,
