+++ packages/web-app/src/index.tsx
@@ -9,6 +9,7 @@ import {WalletMenuProvider} from 'context/walletMenu';
 import {GlobalModalsProvider} from 'context/globalModals';
 import {ApolloClientProvider} from 'context/apolloClient';
 import {ProvidersProvider} from 'context/providers';
+import {TransactionsProvider} from 'context/transactions';
 import {UseSignerProvider} from 'use-signer';
 import {IProviderOptions} from 'web3modal';
 import WalletConnectProvider from '@walletconnect/web3-provider/dist/umd/index.min.js';
@@ -31,11 +32,13 @@ ReactDOM.render(
           <ProvidersProvider>
             <WalletMenuProvider>
               <GlobalModalsProvider>
-                <Router>
-                  <ApolloClientProvider>
-                    <App />
-                  </ApolloClientProvider>
-                </Router>
+                <TransactionsProvider>
+                  <Router>
+                    <ApolloClientProvider>
+                      <App />
+                    </ApolloClientProvider>
+                  </Router>
+                </TransactionsProvider>
               </GlobalModalsProvider>
             </WalletMenuProvider>
           </ProvidersProvider>
