+++ packages/web-app/src/locales/en/common.json
@@ -262,16 +262,19 @@
     "noAddresses": "No address Found"
   },
   "TransactionModal": {
-    "estimatedFees": "Estimated Gas Fees",
-    "synced": "Synced {{time}} sec ag",
-    "totalCost": "Total Cost",
-    "approveToken": "Approve token",
-    "signDeposit": "Sign Deposit",
     "approveSubtitle": "To sign your first transaction, you must approve Aragon Zaragoza to make a transaction with your token.",
+    "approveToken": "Approve token",
+    "depositTitle": "Sign Deposit",
     "dismiss": "Dismiss",
-    "tryAgain": "Try Again",
     "errorLabel": "Error while confirmation",
-    "successLabel": "Transaction successful"
+    "estimatedFees": "Estimated Gas Fees",
+    "publishDao": "Publish Dao",
+    "publishDaoButtonLabel": "Publish DAO Now",
+    "signDeposit": "Sign Deposit",
+    "successLabel": "Transaction successful signed",
+    "synced": "Synced {{time}} sec ago",
+    "totalCost": "Total Cost",
+    "tryAgain": "Try Again"
   },
   "AddActionModal": {
     "title": "Add Action",
