+++ packages/web-app/src/pages/home.tsx
@@ -4,11 +4,9 @@ import {ButtonText} from '@aragon/ui-components';
 import {useNavigate} from 'react-router-dom';
 import {useTranslation} from 'react-i18next';
 import {withTransaction} from '@elastic/apm-rum-react';
-import {useGlobalModalContext} from 'context/globalModals';
 
 const Home: React.FC = () => {
   const {t} = useTranslation();
-  const {open} = useGlobalModalContext();
   const navigate = useNavigate();
 
   return (
@@ -24,12 +22,6 @@ const Home: React.FC = () => {
         className="mx-auto"
         size="large"
         onClick={() => navigate('/create-dao')}
-      />
-      <ButtonText
-        label="Open Transaction Modal"
-        className="mx-auto mt-3"
-        size="large"
-        onClick={() => open('transaction')}
       />
     </div>
   );
